Calculating ductile viscosity on markers + various speedups and bug fixes 
• Introduced the possibility to weaken friction linearly with accumulated
plastic strain (just like cohesion): just specify MAT().mumin for each phase to implement a weakened friction

• Important changes in SiStER_VEP_Rheology:
The calculation of the "ductile viscosity" should be based only on the
viscous component of strain rate, if elasticity is present (i.e., the
elastic strain rate should not be included). The best way to do this is
to use the stress-based formulas for creep viscosity (see Gerya 2010,
p. 189). But these were hard to get to converge, so I introduced a
bisection algorithm (SiStER_viscosity_stress_bisection.m) to determine
a deviatoric stress sII that is most consistent with the desired
viscosity.
If elasticity is off, then the ductile viscosity calculation makes use
of the total strain rate, which is by definition entirely viscous in
this case.

• Caught a bug in SiStER_update_ep.m: the upd