Enabling generalized phase mixing for viscosity
When calculating ductile viscosity (in SiStER_get_ductile_rheology_on_nodes_from_xx.m), the relative fraction of each phase present around any given node is used to compute the viscosity of the mixture. This means we can now "mix" the viscosities of >2 phases that are not consecutively numbered (e.g., a node surrounded by 50% phase 1 and 50% phase 3 will NOT be assigned the viscosity of phase 2, but a mix of 50% viscosity of 1 + 50% viscosity of 2). This involves changes to SiStER_VEP_rheology.m, _get_ductile_rheology_on_nodes_from_xx.m, and SiStER_VEP_rheology.m